depthMap improved

+++ examples/depth_map/depth_linear.frag
@@ -1,8 +1,10 @@
-precision mediump float;
+precision highp float;
 
+// Warning: This shader is used to display the z-Buffer in a color meaningful way
+// To actually implement effects based on the z-Buffer use the other shader versions
 // The fragment eye depth is obtained by solving z_e in terms z_n from here:
-// http://visualcomputing.github.io/Transformations/#/6/14 and:
-// http://visualcomputing.github.io/Transformations/#/6/15
+// http://visualcomputing.github.io/Transformations/#/7/14 and:
+// http://visualcomputing.github.io/Transformations/#/7/15
 // yielding to: z_e = (2 * near * far) / (z_n * (far - near) - far - near) (eq1)
 
 uniform float near;
